+++ src/main/java/figures/interfaces/Colorable.java
@@ -5,5 +5,5 @@ import javafx.scene.paint.Color;
 public interface Colorable {
     void setFigureColor(Color color);
     void setBorderColor(Color color);
-    void setThickness(int thickness);
+    void setThickness(double thickness);
 }
